Final Commit Push
Added more styles rules in the CSS Stylesheet. We added a certain style rule that a certain time in the video it rotates when the guy in the video is dancing on the ceiling the video turns when he's on the ceiling. We added a style rule for the label element so you can click the the button to stop the video and rotation or to begin the video or rotation.